--- src/providers/dataProvider/platforms.ts
@@ -1,19 +1,0 @@
-const platforms: Platform[] = [
-  {
-    id: 1,
-    name: 'plat-1',
-    active: true
-  },
-  {
-    id: 2,
-    name: 'plat-2',
-    active: false
-  },
-  {
-    id: 3,
-    name: 'plat-3',
-    active: true
-  }
-]
-
-export default platforms
